Backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
asyncpg==0.30.0
bcrypt==3.2.2
certifi==2024.12.14
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
decorator==5.2.1
distro==1.9.0
Django==5.1.2
django-appconf==1.0.6
django-ckeditor-5==0.2.15
django-cors-headers==4.6.0
django-js-asset==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
fastapi==0.115.6
fpdf==1.7.2
git-filter-repo==2.47.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
jiter==0.8.2
jwcrypto==1.5.6
jwt==1.3.1
lxml==5.3.1
moviepy==2.1.2
numpy==2.2.3
openai==1.59.8
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
passlib==1.7.4
pillow==10.4.0
proglog==0.1.11
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pydub==0.25.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-jwt==4.1.0
python-multipart==0.0.20
pytz==2025.1
RapidFuzz==3.13.0
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlparse==0.5.1
starlette==0.41.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.34.0
whitenoise==6.7.0
